requirements/base.txt
# Python 3.11.4
# pip3 install django
asgiref==3.7.2
Django==4.2.2
sqlparse==0.4.4

# pip3 install django-environ
django-environ==0.10.0

# python -m pip3 install django-cors-headers
django-simple-history==3.3.0

# pip3 install django-admin-interface
django_admin_interface==0.26.0
django_colorfield==0.9.0
python_slugify==8.0.1
Pillow==9.5.0
text_unidecode==1.3

# pip3 install psycopg2-binary
psycopg2_binary==2.9.6

# pip3 install djangorestframework
djangorestframework==3.14.0
pytz==2023.3

# pip3 install djangorestframework-simplejwt
djangorestframework_simplejwt==5.2.2
PyJWT==2.7.0

# python -m pip install django-cors-headers
django_cors_headers==4.1.0
